feat(compiler): Add support for multiple cases in switch control flow
Currently, the @switch/@case syntax is fairly limited, as cases may only have one guard. This often requires to either copy paste the body of the case-block or to refactor code to use @if / @else if syntax instead. This adds support for the highly requested feature of being able to define multiple guards to a single @case using the syntax @case(a; b; c)